# api/docs/swagger.yaml
definitions:
  models.Author:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.AuthorCU:
    properties:
      name:
        type: string
    type: object
  models.Book:
    properties:
      author_id:
        $ref: '#/definitions/models.Author'
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      category_id:
        items:
          type: string
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      update_at:
        type: string
    type: object
  models.BookCU:
    properties:
      author_id:
        type: string
      category_id:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
info:
  contact: {}
paths:
  /v1/authors/:
    get:
      consumes:
      - application/json
      description: This API for getting list of authors
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Author'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListAuthors
      tags:
      - author
    post:
      consumes:
      - application/json
      description: This API for creating a new author
      parameters:
      - description: AuthorCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuthorCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateAuthor
      tags:
      - author
  /v1/authors/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting author
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteAuthor
      tags:
      - author
    get:
      consumes:
      - application/json
      description: This API for getting author detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetAuthor
      tags:
      - author
    put:
      consumes:
      - application/json
      description: This API for updating author
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: authorUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuthorCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateAuthor
      tags:
      - author
  /v1/books/:
    get:
      consumes:
      - application/json
      description: This API for getting list of books
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      - description: authors
        in: query
        name: authors
        type: string
      - description: categories
        in: query
        name: categories
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: ListBooks
      tags:
      - book
    post:
      consumes:
      - application/json
      description: This API for creating a new book
      parameters:
      - description: BookCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BookCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: CreateBook
      tags:
      - book
  /v1/books/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: DeleteBook
      tags:
      - book
    get:
      consumes:
      - application/json
      description: This API for getting book detail
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: GetBook
      tags:
      - book
    put:
      consumes:
      - application/json
      description: This API for updating book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: bookUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BookCU'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      summary: UpdateBook
      tags:
      - book
...
swagger: "2.0"